requirements.txt
ccxt>=4.2.17
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.5.0
python-dotenv>=1.0.0
ta>=0.10.2
pytest>=7.0.0
PyYAML>=6.0
websocket-client>=1.2.3
requests>=2.27.1
streamlit>=1.32.0
plotly>=5.19.0
psutil>=7.0.0 